requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
requests==2.32.3
pytest==8.2.2
line-bot-sdk